refactor: common metrics fixtures
